Fixed precition of clones. now moves exactly like player. Fixes #10
Fixed by changing all fixedUpdate calls to Update calls, since it increased precision by a large margin.
However, since it still was not perfect, we now save position and velocity data from the player as input events each 0.1 second (as a initial value, can easily be changed later if needed.)
This makes the clone move in the exact same paths as the player and will always end up in the same end position.