Clones can now not phase through walls or eachother. Fixes #17
Clones can no longer phase through something that is not possible anymore.
The issue is fixed by saving the current state for each poll on the player.
If the state on the clone does not match up with the saved state of the player (e.g. if clone collides when player did not), syncing will be disabled for the current run.